requirements.txt
Django==5.0.6
requests==2.31.0
argon2-cffi==23.1.0
python-magic==0.4.27
